Major Update, See Notes
1. "imgloras" are now "imgtags".
I apologize for any inconvenience caused by this change, but it was not sensible to refer to this feature as "imgloras"
This change affects a number of elements so if you are going to update your own files, please carefully refer to the changed lines and filenames.

2. Added support for a really cool extension, 'sd-webui-loractl' (https://github.com/cheald/sd-webui-loractl).
The extension enhances the functionality of A1111s handling of LORAs so that they can be scaled during image generation.
The way this has been incorporated into the bot, is that you may predefine scaling patterns that will be applied sequentially to detected LORAs in the prompt.  Refer to the comments in config.py and try it out for yourself, it can really improve results when processing multiple LORAs!!