Major Update: Tags Feature
First there was a basic "presets" feature.  Then there was "ImgLoras", which became "ImgTags". Meanwhile, settings for new features have been piling up in config.py.

Today I unveil a highly flexible and versatile 'Tags' system, which is a simple way to control most of the bot features while adding an extra layer of customization - the ability to have their effects apply globally, for specific imgmodels, or for specific characters.

With the tags system, there are seemingly limitless possibilities when it comes to triggering: modifying user prompt/payload to the LLM, LLM's prompt/payload for A1111, swapping characters, activating extensions, face swapping, censoring, and so on.